requirements.txt
aio-pika==6.6.0
aiormq==3.2.1
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
multidict==4.7.5
-e git+git@github.com:IOR88/messaging-with-rabbitmq.git@a8d2e4c430cfe489673d6bc0186eedf82558a772#egg=mwr
pamqp==2.3.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
tornado==6.0.4
vine==1.3.0
yarl==1.4.2
zipp==3.1.0
